added grep to mcp tools and docs
