+++ crates/sem-core/src/utils/mod.rs
@@ -1,1 +1,2 @@
 pub mod hash;
+pub mod paths;
