+++ docs/benchmarks.html
@@ -177,13 +177,13 @@
     <section id="hero">
       <h2>sem-mcp performance</h2>
       <p class="section-desc">
-        6 MCP tools. Entity-level intelligence for agents.<br>
+        7 MCP tools. Entity-level intelligence for agents.<br>
         Real numbers from sem's own codebase (48 Rust files).
       </p>
 
       <div class="speed-row" style="margin-top: 0;">
         <div class="speed-pill">
-          <span class="number">6</span>
+          <span class="number">7</span>
           <span class="label">MCP tools</span>
         </div>
         <div class="speed-pill">
